requirements.txt
Django==3.2.16
djangorestframework==3.14.0
django-webpack-loader==1.6.0
django-treebeard==4.5.1
django-cte==1.2.1

django-storages==1.13.1
# boto3 is part of django-storages, but latest version has incorrect dependencies
boto3==1.24.52
